Expand the Potato genome annotation
We now have access to GO annotation and Gene <-> Protein association
for the Potato genome. This gives us a much larger annotation to play
with and makes it really close to the Tomato annotation wrt the size of
the network considered by M2S for integration.

Fix the urls to point to their right sources.